Split FlameGraphCell into flame-model + FlameGraphScene (#1089)
Extract the pure data/geometry/format model (flame-model.ts) and the
THREE.js renderer behind a createFlameScene(...) handle (FlameGraphScene.ts).
FlameGraphCell becomes a shell that keeps view/interaction state and feeds a
per-frame snapshot to the scene. Repoints the test import and adds
flame-model unit tests (hitTest, spanColor, formatDuration, lane geometry).
